requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scipy>=1.12,<2.0
plotly>=5.20,<6.0
yfinance>=0.2.40,<1.0
